src/docs_stackoverflow_agents/requirements.txt
aiohttp==3.11.11
firecrawl_py==1.11.0
gradio==5.16.1
groq==0.18.0
openai==1.63.2
pydantic_ai==0.0.24
python-dotenv==1.0.1
supabase==2.13.0
tqdm==4.67.1